Added option to show details for specific interface. Also changed logs to color

+++ macos/network_info.sh
@@ -18,9 +18,11 @@ dir_name=`dirname $0`
 my_path=$(cd $dir_name; pwd -P)
 
 log_file="/tmp/$(echo $my_name|cut -d. -f1).log"
-options="ewvh?"
+log_init=0
+options="i:ewvh?"
 verbose=0
 ssid=""
+net_dev="en0"
 airport="/System/Library/PrivateFrameworks/Apple80211.framework/Versions/Current/Resources/airport"
 
 # ensure path for cron runs
@@ -30,10 +32,11 @@ usage() {
   cat << EOF
 
   Usage: $my_name [options]
-     -e         ---> print IP, MAC ... etc of a ethernet (wired) interface
-     -w         ---> print IP, MAC ... etc of a WiFi (wireless) interface
-     -v         ---> verbose mode prints info messages, otherwise just errors are printed
-     -h         ---> print usage/help
+     -i <interface> ---> show details of specified interface [example: en0, en1...etc]
+     -e             ---> show details of wired interface [on multi-homed, first device shown]
+     -w             ---> show details of wireless interface [on multi-homed, first device shown]
+     -v             ---> verbose mode prints info messages, otherwise just errors are printed
+     -h             ---> print usage/help
 
   example: $my_name -w
   
@@ -41,47 +44,66 @@ EOF
   exit 0
 }
 
-write_log() {
-  local msg_type=$1
-  local msg=$2
+log.init() {
+  if [ $log_init -eq 1 ] ; then
+    return
+  fi
 
-  case $msg_type in
-    # skip info if verbose is not set [default]
-    info|INFO)
-      if [ $verbose -eq 0 ] ; then
-        return;
-      fi
-      echo -e "\e[1;36m[INFO]\e[0m $msg" | tee -a $log_file      
-      ;;
-    stat|STAT)
-      echo -e "\e[1;34m[STAT]\e[0m $msg" | tee -a $log_file
-      ;;
-    warn|WARN)
-      echo -e "\e[1;33m[WARN]\e[0m $msg" | tee -a $log_file
-      ;;
-    error|ERROR)
-      echo -e "\e[1;31m[ERROR]\e[0m $msg" | tee -a $log_file
-      ;;
-  esac
-}
-init_log() {
+  log_init=1
   if [ -f $log_file ] ; then
     rm -f $log_file
   fi
-  write_log "stat" "$my_version"
-  write_log "info" "Running from: $my_path"
-  write_log "info" "Start time:   `date +'%m/%d/%y %r'` ..."
+  echo -e "\e[0;34m$my_version, `date +'%m/%d/%y %r'` \e[0m" | tee -a $log_file
+}
+
+log.info() {
+  if [ $verbose -eq 0 ] ; then
+    return;
+  fi
+  log.init
+  local msg=$1
+  echo -e "\e[0;32m$msg\e[0m" | tee -a $log_file 
+}
+log.debug() {
+  if [ $verbose -eq 0 ] ; then
+    return;
+  fi
+  log.init
+  local msg=$1
+  echo -e "\e[1;30m$msg\e[0m" | tee -a $log_file 
+}
+
+log.stat() {
+  log.init
+  local msg=$1
+  echo -e "\e[0;34m$msg\e[0m" | tee -a $log_file 
+}
+
+log.warn() {
+  log.init
+  local msg=$1
+  echo -e "\e[1;33m$msg\e[0m" | tee -a $log_file 
+}
+log.error() {
+  log.init
+  local msg=$1
+  echo -e "\e[1;31m$msg\e[0m" | tee -a $log_file 
+}
+
+get_interface() {
+  local dev_string=$1
+  net_dev=$(networksetup -listallhardwareports| awk "/Hardware Port: ${dev_string}/ {getline; print \$2}")
+  if [ -z "$net_dev" ] ; then
+    log.error "Not able to determine your network interface!"
+    exit 2
+  fi
+  log.info "The network interface is '$net_dev'"
 }
 
 print_info() {
   local dev_string=$1
-  local net_dev=$(networksetup -listallhardwareports| awk "/Hardware Port: ${dev_string}/ {getline; print \$2}")
-  if [ -z "$net_dev" ] ; then
-    write_log "error" "Not able to determine your network interface!"
-    exit 2
-  fi
-  write_log "info" "The network interface is '$net_dev'"
   lease_secs=`ipconfig getoption $net_dev lease_time`
+  log.stat "Total number of interfaces available: `ipconfig ifcount`"
   # read IP, MAC ... etc
   echo "    Interface:    $net_dev"
   echo "    Type:         $dev_string"
@@ -98,23 +120,23 @@ print_info() {
   if [ "$dev_string" = "Wi-Fi" ] ; then
     rssi=`$airport -I |awk '/ agrCtlRSSI:/ {print $2}'`
     noise=`$airport -I |awk '/ agrCtlNoise:/ {print $2}'`
-  echo "    WiFi Name:         `$airport -I |awk '/ SSID:/ {print $2}'`"
-  echo "    WiFi Channel:      `$airport -I |awk '/ channel:/ {print $2}'`"
-  echo "    WiFi Auth:         `$airport -I |awk '/ link auth:/ {print $3}'`"
-  echo "    WiFi Channel:      `$airport -I |awk '/ channel:/ {print $2}'`"
-  echo "    WiFI RSSI:         $rssi [range: (-100,0) note: closer to 0 is better, ex: -55 is pretty damn good]"
-  echo "    WiFi Noise:        $noise [range: (-120,0) note: closer to -120 is better]"
-  echo "    WiFi Quality:      $((rssi - noise)) [should be above 20]"
-  echo "    WiFi Last TxnRate: `$airport -I |awk '/ lastTxRate:/ {print $2}'` mbps"
+    echo "    WiFi Name:         `$airport -I |awk '/ SSID:/ {print $2}'`"
+    echo "    WiFi Channel:      `$airport -I |awk '/ channel:/ {print $2}'`"
+    echo "    WiFi Auth:         `$airport -I |awk '/ link auth:/ {print $3}'`"
+    echo "    WiFi Channel:      `$airport -I |awk '/ channel:/ {print $2}'`"
+    echo "    WiFI RSSI:         $rssi [range: (-100,0) note: closer to 0 is better, ex: -55 is pretty damn good]"
+    echo "    WiFi Noise:        $noise [range: (-120,0) note: closer to -120 is better]"
+    echo "    WiFi Quality:      $((rssi - noise)) [should be at least 20 or greater]"
+    echo "    WiFi Last TxnRate: `$airport -I |awk '/ lastTxRate:/ {print $2}'` mbps"
   fi
 
   exit 0
 }
 
 # ----------  main --------------
-init_log
+log.init
 if [ "$os_name" != "Darwin" ] ; then
-  write_log "error" "This script is intended to run on MacOS!"
+  log.error "This script is intended to run on MacOS!"
   exit 1
 fi
 
@@ -122,10 +144,22 @@ fi
 while getopts $options opt ; do
   case $opt in
     w)
+      get_interface "Wi-Fi"
       print_info "Wi-Fi"
       ;;
     e)
+      get_interface "Ethernet"
       print_info "Ethernet"
+      ;;
+    i)
+      net_dev="$OPTARG"
+      # see if this is a wired or wireless
+      networksetup -getairportnetwork $net_dev 2>&1 >/dev/null
+      if [ $? -eq 0 ] ; then
+        print_info "Wi-Fi"
+      else
+        print_info "Ethernet/Other"
+      fi
       ;;
     v)
       verbose=1
@@ -137,5 +171,5 @@ while getopts $options opt ; do
 done
 
 # if no argument is provided, just show wifi
+get_interface "Wi-Fi"
 print_info "Wi-Fi"
-
